fix: onboarding developer mode trigger and screen
