+++ app/src/screens/Preface.tsx
@@ -3,37 +3,56 @@ import {
   useStore,
   testIdWithKey,
   DispatchAction,
-  AuthenticateStackParams,
   Screens,
   CheckBoxRow,
   Button,
   ButtonType,
   Link,
+  OnboardingStackParams,
 } from '@hyperledger/aries-bifold-core'
 import { useNavigation } from '@react-navigation/native'
 import { StackNavigationProp } from '@react-navigation/stack'
-import React, { useState } from 'react'
+import React, { useRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { Linking, ScrollView, StyleSheet, Text, View } from 'react-native'
+import { Linking, Pressable, ScrollView, StyleSheet, Text, View } from 'react-native'
 import { SafeAreaView } from 'react-native-safe-area-context'
 
 const Preface: React.FC = () => {
   const [, dispatch] = useStore()
   const [checked, setChecked] = useState(false)
   const { t } = useTranslation()
-  const navigation = useNavigation<StackNavigationProp<AuthenticateStackParams>>()
+  const developerOptionCount = useRef(0)
+  const touchCountToEnableBiometrics = 9
+  const navigation = useNavigation<StackNavigationProp<OnboardingStackParams>>()
   const { Assets, OnboardingTheme, TextTheme } = useTheme()
+
   const onSubmitPressed = () => {
     dispatch({
       type: DispatchAction.DID_SEE_PREFACE,
     })
     navigation.navigate(Screens.Onboarding)
   }
+
   const onPressInfoLink = () => {
     Linking.openURL('https://www2.gov.bc.ca/gov/content/governments/government-id/bc-wallet#where')
   }
+
   const onPressShowcaseLink = () => {
     Linking.openURL('https://digital.gov.bc.ca/digital-trust/showcase/')
+  }
+
+  const incrementDeveloperMenuCounter = () => {
+    if (developerOptionCount.current >= touchCountToEnableBiometrics) {
+      developerOptionCount.current = 0
+      dispatch({
+        type: DispatchAction.ENABLE_DEVELOPER_MODE,
+        payload: [true],
+      })
+      navigation.navigate(Screens.Developer)
+      return
+    }
+
+    developerOptionCount.current = developerOptionCount.current + 1
   }
 
   const style = StyleSheet.create({
@@ -55,7 +74,9 @@ const Preface: React.FC = () => {
         <View style={style.screenContainer}>
           <View style={style.contentContainer}>
             <Assets.svg.preface style={{ alignSelf: 'center', marginBottom: 20 }} height={200} />
-            <Text style={[TextTheme.headingTwo]}>{t('Preface.PrimaryHeading')}</Text>
+            <Pressable onPress={incrementDeveloperMenuCounter} testID={testIdWithKey('DeveloperCounter')}>
+              <Text style={[TextTheme.headingTwo]}>{t('Preface.PrimaryHeading')}</Text>
+            </Pressable>
             <Text style={[TextTheme.normal, { marginTop: 10, marginBottom: 10 }]}>{t('Preface.Paragraph1')}</Text>
             <Link style={{ marginTop: 10, marginBottom: 10 }} onPress={onPressInfoLink} linkText={t('Preface.Link1')} />
             <Text style={[TextTheme.normal, { marginTop: 10 }]}>{t('Preface.Paragraph2')}</Text>
